Update tests: - update pom.xml file - separate parameterized tests - add jococo report
