+++ pom.xml
@@ -64,7 +64,13 @@
                             <goal>prepare-agent</goal>
                         </goals>
                     </execution>
-
+                    <execution>
+                        <id>report</id>
+                        <phase>verify</phase>
+                        <goals>
+                            <goal>report</goal>
+                        </goals>
+                    </execution>
                 </executions>
             </plugin>
         </plugins>
