+++ src/test/java/CatTest.java
@@ -1,13 +1,9 @@
-import com.example.Animal;
 import com.example.Cat;
 import com.example.Feline;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-
 import org.mockito.junit.MockitoJUnitRunner;
-
-
 import static org.junit.Assert.assertEquals;
 
 @RunWith(MockitoJUnitRunner.class)
