+++ src/test/java/FelineTest.java
@@ -1,29 +1,11 @@
 import com.example.Feline;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
+
 import static org.junit.Assert.assertEquals;
 
-@RunWith(Parameterized.class)
 public class FelineTest {
 
     Feline feline = new Feline();
-    private final int actualKittensCount;
-    private final int expectedKittensCount;
-
-    public FelineTest(int actualKittensCount, int expectedKittensCount) {
-        this.actualKittensCount = actualKittensCount;
-        this.expectedKittensCount = expectedKittensCount;
-    }
-
-    @Parameterized.Parameters // добавили аннотацию
-    public static Object[][] getKittensCount() {
-        return new Object[][]{
-                {1, 1},
-                {4, 4},
-                {2, 2},
-        };
-    }
 
     @Test
     public void getFelineFamilyTest() {
@@ -40,14 +22,7 @@ public class FelineTest {
     }
 
     @Test
-    public void getFelineKittensWithParamsTest() {
-        int actualCount = feline.getKittens(actualKittensCount);
-        int expectedCount = expectedKittensCount;
-        assertEquals(expectedCount, expectedCount);
-    }
-
-    @Test
-    public void getFelineMeatTest (){
+    public void getFelineMeatTest() {
 
     }
 }
